rag_processor/requirements.txt
# rag_processor/requirements.txt
kafka-python
chromadb
google-generativeai
pysqlite3-binary
